@@ -1,7 +1,9 @@
 {
   "name": "fact-service",
-  "version": "1.1.1",
-  "description": "Enhanced Key-Value lookup service. Uses Nodejs, TypeScript, Express & Knex Project.",
+  "version": "1.4.0",
+  "type": "module",
+  "description": "Starter Key-Value Platform + API. Feat. Nodejs, TypeScript, Express & Knex Project. Supports Postgres, Redis, DynamoDB, Firestore, Cassandra & more.",
+  
   "main": "index.js",
   "bin": {
     "fact-cli": "./bin/fact-cli"
@@ -12,7 +14,8 @@
   },
   "homepage": "https://github.com/justsml/fact-service",
   "scripts": {
-    "start": "nodemon --watch './**/*.ts' --exec 'ts-node' index.ts",
+    "start": "nodemon --watch './**/*.ts' --exec 'ts-node' ./servers/express/index.ts",
+    "start:elysia": "bun run --watch ./servers/elysia/index.ts",
     "test": "vitest",
     "test:pg": "DB_ADAPTER=postgres vitest",
     "test:redis": "DB_ADAPTER=redis vitest",
@@ -34,28 +37,40 @@
     "nodejs",
     "typescript",
     "express",
-    "knex"
+    "knex",
+    "postgres",
+    "postgresql",
+    "redis",
+    "dynamodb",
+    "cassandra",
+    "firestore",
+    "api"
   ],
   "author": {
     "name": "Dan Levy",
     "url": "https://danlevy.net"
   },
   "engines": {
-    "node": ">=16.0.0",
+    "node": ">=16.14.0",
     "npm": ">=7.0.0",
     "yarn": ">=1.22.0"
   },
   "engineStrict": true,
   "license": "MIT",
   "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.478.0",
-    "@aws-sdk/lib-dynamodb": "^3.478.0",
+    "@aws-sdk/client-dynamodb": "^3.535.0",
+    "@aws-sdk/lib-dynamodb": "^3.535.0",
+    "@bogeychan/elysia-logger": "^0.0.18",
+    "@elysiajs/bearer": "^1.0.0",
+    "@elysiajs/cors": "^1.0.0",
+    "@elysiajs/swagger": "^1.0.0",
     "@google-cloud/firestore": "^7.1.0",
-    "@elite-libs/auto-config": "^1.4.1",
-    "axios": "^1.6.2",
+    "axios": "^1.6.8",
     "cassandra-driver": "^4.7.2",
     "cors": "^2.8.5",
     "dotenv": "^16.3.1",
+    "elysia": "^1.0.1",
+    "elysia-helmet": "^1.0.2",
     "express": "^4.18.2",
     "helmet": "^7.1.0",
     "ioredis": "^5.3.2",
@@ -63,26 +78,30 @@
     "lodash": "^4.17.21",
     "morgan": "^1.10.0",
     "pg": "^8.11.3",
-    "pino": "^8.17.1",
-    "pino-http": "^8.6.0",
-    "pino-pretty": "^10.3.0",
-    "ulidx": "^2.2.1"
+    "pino": "^8.19.0",
+    "pino-http": "^9.0.0",
+    "pino-pretty": "^10.3.1",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@faker-js/faker": "^8.3.1",
-    "@types/cors": "^2.8.12",
-    "@types/express": "^4.17.14",
-    "@types/lodash": "^4.14.189",
-    "@types/morgan": "^1.9.3",
-    "@types/ms": "^0.7.31",
-    "@types/node": "^18.11.8",
-    "@types/supertest": "^2.0.16",
+    "@faker-js/faker": "^8.4.1",
+    "@types/bun": "^1.0.8",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/lodash": "^4.17.0",
+    "@types/morgan": "^1.9.9",
+    "@types/ms": "^0.7.34",
+    "@types/node": "^20.11.28",
+    "@types/supertest": "^6.0.2",
+    "@types/yargs": "^17.0.32",
+    "bun-types": "^1.0.31",
     "ms": "^2.1.3",
-    "nodemon": "^3.0.2",
-    "prettier": "^3.1.1",
-    "supertest": "^6.3.3",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.3.3",
-    "vitest": "^1.1.0"
+    "nodemon": "^3.1.0",
+    "prettier": "^3.2.5",
+    "supertest": "^6.3.4",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.2",
+    "vite-tsconfig-paths": "^4.3.2",
+    "vitest": "^1.4.0"
   }
 }
