@@ -1,10 +1,10 @@
 import axios from "axios";
 // import { toArray } from "../../common/arrayUtils";
-import type { Fact, FactAdapter } from "./types";
+import type { FactAdapter, FactEntity } from "./types";
 
 const FactsConfig = {
-  apiUrl: process.env.API_URL ?? "http://localhost:3000/api/facts",
-  apiToken: process.env.API_TOKEN ?? null,
+  apiUrl: process.env["API_URL"] ?? "http://localhost:3000/api",
+  apiToken: process.env["API_TOKEN"] ?? null,
 };
 
 const client = axios.create({
@@ -22,23 +22,26 @@ const enc = encodeURIComponent;
 const FactApiClient: FactAdapter = {
   _name: "http",
 
-  set: (fact) => client.put(`/`, fact).then((res) => res.data),
+  set: ({ key, ...fact }) =>
+    client.put(`/facts/${enc(key)}`, fact).then((res) => res.data),
 
   get: ({ key }) =>
-    client.get<Fact>(`/${encodeURIComponent(key)}`).then((res) => res.data),
+    client.get<FactEntity>(`/facts/${enc(key)}`).then((res) => res.data),
 
   // updateByPathKey: (update, fact) =>
   //   client
   //     .post<Fact[]>(`/${enc(update.path)}/${enc(update.key.toString())}`, fact)
   //     .then((res) => res.data),
 
-  del: ({ key }) => client.delete(`/${key}`).then((res) => res.data),
+  del: ({ key }) => client.delete(`/facts/${enc(key)}`).then((res) => res.data),
 
   // getPathCounts: () =>
   //   client.get<PathCountResults>(`/?count=path`).then((res) => res.data),
 
   find: ({ keyPrefix }) =>
-    client.get<Fact[]>(`/?keyPrefix=${enc(keyPrefix)}`).then((res) => res.data),
+    client
+      .get<FactEntity[]>(`/query/${enc(keyPrefix)}`)
+      .then((res) => res.data),
 
   // findAllFactsByPath: ({ path, limit }) =>
   //   client
