@@ -1,5 +1,6 @@
 import isArray from "lodash/isArray";
 import type { Request } from "express";
+import type EventEmitter from "events";
 
 export const toCsvString = (value: any) =>
   value == null
@@ -10,6 +11,26 @@ export const toCsvString = (value: any) =>
 
 export const getKeyFromParamsOrQuery = (request: Request) => {
   let { key } = request.params;
-  if (!key && request.query?.key) key = `${toCsvString(request.query?.key)}`;
+  if (!key && request.query?.["key"])
+    key = `${toCsvString(request.query?.["key"])}`;
   return { key };
+};
+
+/**
+ * Register an error handler that will exit the process
+ * if the error count exceeds the limit.
+ */
+export const registerErrorHandlerLimit = (
+  events: EventEmitter,
+  message: string = "Error: Check your configuration!",
+  limit: number = 10,
+) => {
+  let errorCount = 0;
+  events.on("error", (error) => {
+    errorCount++;
+    if (errorCount >= limit) {
+      console.error(error, message + " %o", error);
+      process.exit(1);
+    }
+  });
 };
