@@ -1,14 +1,14 @@
-import type { Fact, FactAdapter } from "../../factService/types";
+import { isFactEntity, type FactAdapter } from "../../factService/types";
 import { Firestore } from "@google-cloud/firestore";
 import { logger } from "../../../common/logger";
-import { isArray } from "lodash";
+import isArray from "lodash/isArray";
 
 const firestore = new Firestore({
-  projectId: process.env.FIRESTORE_PROJECT_ID,
+  projectId: process.env["FIRESTORE_PROJECT_ID"],
 });
 
 firestore.settings({
-  host: process.env.FIRESTORE_HOST,
+  host: process.env["FIRESTORE_HOST"],
   ssl: false,
   ignoreUndefinedProperties: true,
 });
@@ -18,43 +18,47 @@ const COLLECTION_NAME = "fact_store";
 export const adapter: FactAdapter = {
   _name: "firestore",
 
-  set: async ({ key, fact }) => {
+  async set({ key, value }) {
     try {
       const docRef = firestore.collection(COLLECTION_NAME).doc(key);
-      await docRef.set(fact);
-      return { key, ...fact };
+      if (typeof value === "string" || typeof value === "number" || typeof value === "boolean") throw new Error("Invalid primitive value");
+
+      await docRef.set(value);
+      return { key, value };
     } catch (error) {
       logger.error("ERROR %o", error);
       throw error;
     }
   },
 
-  get: async ({ key }) => {
+  async get({ key }) {
     try {
       const docRef = firestore.collection(COLLECTION_NAME).doc(key);
       const doc = await docRef.get();
       if (!doc.exists) {
         throw new Error("Document not found");
       }
-      return { key, ...doc.data() }; // Adjust as needed
+      const result = { key, ...doc.data() }; // Adjust as needed
+      if (isFactEntity(result)) return result;
+      throw new Error("Invalid fact");
     } catch (error) {
-      logger.error("ERROR %o", error);
+      logger.error(`ERROR[${this._name}] %o`, error);
       throw error;
     }
   },
 
-  del: async ({ key }) => {
+  async del({ key }) {
+    let _key = isArray(key) ? key : [key];
     try {
       let count = 0;
-      key = isArray(key) ? key : [key];
-      for await (const k of key) {
+      for await (const k of _key) {
         await firestore.collection(COLLECTION_NAME).doc(k).delete();
         count++;
       }
       return {
         count,
         success: true,
-        message: `Deleted ${count} fact(s) with id(s) ${key.join(", ")}`,
+        message: `Deleted ${count} fact(s) with id(s) ${_key.join(", ")}`,
       };
     } catch (error) {
       logger.error("ERROR %o", error);
@@ -62,7 +66,7 @@ export const adapter: FactAdapter = {
     }
   },
 
-  find: async ({ keyPrefix }) => {
+  async find({ keyPrefix }) {
     try {
       const querySnapshot = await firestore
         .collection(COLLECTION_NAME)
@@ -70,10 +74,29 @@ export const adapter: FactAdapter = {
         .where("key", "<", keyPrefix + "\uf8ff")
         .get();
 
-      return querySnapshot.docs.map((doc) => ({
+      const results = querySnapshot.docs.map((doc) => ({
         key: doc.id,
         ...doc.data(),
-      })) as Fact[]; // Adjust as needed
+      }));
+
+      const validResults = results.filter(isFactEntity);
+
+      const totalKeysFound = results.length;
+      const validKeysFound = validResults.length;
+
+      logger.debug("Found %d keys", validResults.length);
+      logger.info(
+        {
+          query: keyPrefix,
+          validKeysFound,
+          totalKeysFound,
+        },
+        this._name,
+      );
+      if (totalKeysFound > validKeysFound)
+        throw Error(`Invalid fact(s) found: ${totalKeysFound - validKeysFound}`);
+
+      return validResults;
     } catch (error) {
       logger.error("ERROR %o", error);
       throw error;
