@@ -5,7 +5,7 @@ import { getSetup } from "../lib/providers";
 
 (async () => {
   const results = Promise.allSettled(
-    testAdapters.map(async (adapter) => {
+    testAdapters.map(async (adapter: string) => {
       const setupFn = getSetup(adapter);
       const result = await setupFn();
       return [adapter, result];
